Add the individual
