+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/ConsumerImpl.java
@@ -49,7 +49,6 @@ import java.util.TreeMap;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ScheduledExecutorService;
@@ -61,7 +60,6 @@ import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.stream.Collectors;
 
-import io.netty.util.TimerTask;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.pulsar.client.api.Consumer;
 import org.apache.pulsar.client.api.ConsumerCryptoFailureAction;
@@ -83,6 +81,7 @@ import org.apache.pulsar.client.api.transaction.TxnID;
 import org.apache.pulsar.client.impl.conf.ConsumerConfigurationData;
 import org.apache.pulsar.client.impl.crypto.MessageCryptoBc;
 import org.apache.pulsar.client.impl.transaction.TransactionImpl;
+import org.apache.pulsar.client.impl.AckTimeoutTracker.OpForAckCallBack;
 import org.apache.pulsar.client.util.RetryMessageUtil;
 import org.apache.pulsar.common.protocol.Commands;
 import org.apache.pulsar.common.api.EncryptionContext;
@@ -105,7 +104,6 @@ import org.apache.pulsar.common.schema.SchemaType;
 import org.apache.pulsar.common.util.FutureUtil;
 import org.apache.pulsar.common.util.SafeCollectionUtils;
 import org.apache.pulsar.common.util.collections.BitSetRecyclable;
-import org.apache.pulsar.common.util.collections.ConcurrentLongHashMap;
 import org.apache.pulsar.common.util.collections.ConcurrentOpenHashMap;
 import org.apache.pulsar.common.util.collections.GrowableArrayBlockingQueue;
 import org.apache.pulsar.transaction.common.exception.TransactionConflictException;
@@ -190,12 +188,9 @@ public class ConsumerImpl<T> extends ConsumerBase<T> implements ConnectionHandle
 
     private final boolean createTopicIfDoesNotExist;
     //pending ack
-    private final ConcurrentLongHashMap<OpForAckCallBack> ackRequests =
-            new ConcurrentLongHashMap<>(16, 1);
-    private final ConcurrentLinkedQueue<RequestTime> ackTimeoutQueue;
-    private volatile Timeout ackTimeTracker;
+    private final AckTimeoutTracker<T> ackTimeTracker;
 
-    private static class RequestTime {
+    protected static class RequestTime {
         final long creationTimeMs;
         final long requestId;
 
@@ -345,9 +340,7 @@ public class ConsumerImpl<T> extends ConsumerBase<T> implements ConnectionHandle
         }
 
         topicNameWithoutPartition = topicName.getPartitionedTopicName();
-        this.ackTimeoutQueue = new ConcurrentLinkedQueue<>();
-        this.ackTimeTracker = client.timer().newTimeout(new AckTimeoutTimer(),
-                conf.getAckResponseTimeout(), TimeUnit.MILLISECONDS);
+        this.ackTimeTracker = new AckTimeoutTracker(this);
         grabCnx();
     }
 
@@ -2387,102 +2380,16 @@ public class ConsumerImpl<T> extends ConsumerBase<T> implements ConnectionHandle
                 validationError, properties, txnID.getMostSigBits(), txnID.getLeastSigBits(), requestId);
         OpForAckCallBack op = OpForAckCallBack.create(cmd, callBack, messageId,
                 new TxnID(txnID.getMostSigBits(), txnID.getLeastSigBits()));
-        ackRequests.put(requestId, op);
-        ackTimeoutQueue.add(new RequestTime(System.currentTimeMillis(), requestId));
+        ackTimeTracker.add(op, requestId);
         unAckedMessageTracker.remove(messageId);
         cmd.retain();
         cnx().ctx().writeAndFlush(cmd, cnx().ctx().voidPromise());
         return callBack;
     }
 
-    private static class OpForAckCallBack {
-        protected ByteBuf cmd;
-        protected CompletableFuture<Void> callback;
-        protected MessageIdImpl messageId;
-        protected TxnID txnID;
-
-        static OpForAckCallBack create(ByteBuf cmd, CompletableFuture<Void> callback, MessageId messageId, TxnID txnID) {
-            OpForAckCallBack op = RECYCLER.get();
-            op.callback = callback;
-            op.cmd = cmd;
-            op.messageId = (MessageIdImpl) messageId;
-            op.txnID = txnID;
-            return op;
-        }
-        private OpForAckCallBack(Recycler.Handle<OpForAckCallBack> recyclerHandle) {
-            this.recyclerHandle = recyclerHandle;
-        }
-
-        void recycle() {
-            recyclerHandle.recycle(this);
-        }
-
-        private final Recycler.Handle<OpForAckCallBack> recyclerHandle;
-        private static final Recycler<OpForAckCallBack> RECYCLER = new Recycler<OpForAckCallBack>() {
-            @Override
-            protected OpForAckCallBack newObject(Handle<OpForAckCallBack> handle) {
-                return new OpForAckCallBack(handle);
-            }
-        };
-    }
-
-    private class AckTimeoutTimer implements TimerTask {
-        @Override
-        public void run(Timeout timeout) throws Exception {
-            if (timeout.isCancelled()) {
-                return;
-            }
-            long timeToWaitMs;
-            if (getState() == State.Closing || getState() == State.Closed) {
-                return;
-            }
-            RequestTime peeked = ackTimeoutQueue.peek();
-            while (peeked != null && peeked.creationTimeMs + conf.getAckResponseTimeout()
-                    - System.currentTimeMillis() <= 0) {
-                RequestTime lastPolled = ackTimeoutQueue.poll();
-                if (lastPolled != null) {
-                    OpForAckCallBack op = ackRequests.remove(lastPolled.requestId);
-                    if (op != null && !op.callback.isDone()) {
-                        StringBuilder stringBuilder = new StringBuilder();
-                        stringBuilder.append("MessageId : [")
-                                .append(op.messageId)
-                                .append("]Could not get response from broker within given timeout!");
-                        if (op.txnID.getLeastSigBits() != -1 && op.txnID.getMostSigBits() != -1) {
-                            stringBuilder.append(" TxnId : ")
-                                    .append(op.txnID);
-                        }
-                        op.callback.completeExceptionally(new PulsarClientException
-                                .TimeoutException(stringBuilder.toString()));
-                        if (log.isDebugEnabled()) {
-                            log.debug("MessageId : [{}] Ack RequestId : {} TxnId : [{}] is timeout.",
-                                    op.messageId, lastPolled.requestId, op.txnID);
-                        }
-                        onTransactionAckResponse(op);
-                    }
-                } else {
-                    break;
-                }
-                peeked = ackTimeoutQueue.peek();
-            }
-
-            if (peeked == null) {
-                timeToWaitMs = conf.getAckResponseTimeout();
-            } else {
-                long diff = (peeked.creationTimeMs + client.getConfiguration()
-                        .getOperationTimeoutMs()) - System.currentTimeMillis();
-                if (diff <= 0) {
-                    timeToWaitMs = conf.getAckResponseTimeout();
-                } else {
-                    timeToWaitMs = diff;
-                }
-            }
-            ackTimeTracker = client.timer().newTimeout(this, timeToWaitMs, TimeUnit.MILLISECONDS);
-        }
-    }
-
     protected void ackReceipt(PulsarApi.CommandAckReceipt commandAckReceipt) {
         checkArgument(commandAckReceipt.getRequestId() >= 0);
-        OpForAckCallBack callBackOp = ackRequests.remove(commandAckReceipt.getRequestId());
+        OpForAckCallBack callBackOp = ackTimeTracker.remove(commandAckReceipt.getRequestId());
         if (callBackOp == null || callBackOp.callback.isDone()) {
             log.error("Ack request has been handled requestId : {}", commandAckReceipt.getRequestId());
         } else {
@@ -2497,7 +2404,7 @@ public class ConsumerImpl<T> extends ConsumerBase<T> implements ConnectionHandle
 
     protected void ackError(PulsarApi.CommandAckError commandAckError) {
         checkArgument(commandAckError.getRequestId() >= 0);
-        OpForAckCallBack callBackOp = ackRequests.remove(commandAckError.getRequestId());
+        OpForAckCallBack callBackOp = ackTimeTracker.remove(commandAckError.getRequestId());
         if (callBackOp == null || callBackOp.callback.isDone()) {
             log.error("Ack request has been handled requestId : {}", commandAckError.getRequestId());
         } else {
